--- backend/Application/Interfaces/Authentication/IJwtTokenGenerator.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Application.Interfaces.Authentication
-{
-    public interface IJwtTokenGenerator
-    {
-        string GenerateToken(Guid userId, string firstName, string lastName);
-    }
-}
