+++ backend/Application/Services/Authentication/AuthenticationResult.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Domain.Model;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -7,10 +8,7 @@ using System.Threading.Tasks;
 namespace Application.Services.Authentication
 {
     public record AuthenticationResult(
-       Guid Id,
-       string FirstName,
-       string LastName,
-       string Email,
+       User User,
        string Token
     );
 }
