+++ backend/Domain/Interfaces/Repository/IServiceRepository.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.Interfaces
+namespace Domain.Interfaces.Repository
 {
     public interface IServiceRepository
     {
