+++ backend/Domain/Interfaces/Repository/IUserRepository.cs
@@ -1,0 +1,15 @@
+﻿using Domain.Model;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Domain.Interfaces.Repository
+{
+    public interface IUserRepository
+    {
+        User? GetUserByEmail(string email);
+        void Add(User user);
+    }
+}
