+++ backend/Infrastructure/Repositories/ServiceRepository.cs
@@ -1,11 +1,11 @@
 ﻿using Domain.Model;
 using Infrastructure;
-using Domain.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Domain.Interfaces.Repository;
 
 namespace Infrastructure.Repositories
 {
