+++ backend/Infrastructure/Repositories/UserRepository.cs
@@ -1,4 +1,5 @@
-﻿using Domain.Model;
+﻿using Domain.Interfaces.Repository;
+using Domain.Model;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -8,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace Infrastructure.Repositories
 {
-    public class UserRepository : Domain.Interfaces.IUserRepository
+    public class UserRepository : IUserRepository
     {
         private readonly ApplicationDbContext _dbContext;
